chore(tests): move refill settlement tests to test_gas_settlement
